refactor to use support dir
